+++ src/app/components/components.ts
@@ -1,0 +1,6 @@
+export { default as AvatarAppIntro } from "./avatar-app-intro/avatar-app-intro";
+export { default as Footer } from "./footer/footer";
+export { default as Header } from "./header/header";
+export { default as Recipe } from "./recipe/recipe";
+export { default as SearchBar } from "./search-bar/search-bar";
+export { default as ThemeChangerBtn } from "./theme-changer-btn/theme-changer-btn";
